health-check/requirements.txt
connexion[flask]
flask-cors
pyyaml
apscheduler
requests
uvicorn